+++ apps/backend/src/controllers/admin/allQueues.ts
@@ -2,7 +2,6 @@ import {
   bilbomdQueue,
   scoperQueue,
   multimdQueue,
-  pdb2crdQueue,
   deleteBilboMDJobsQueue
 } from '../../queues/index.js'
 import { Queue } from 'bullmq'
@@ -11,6 +10,5 @@ export const allQueues: { [name: string]: Queue } = {
   bilbomd: bilbomdQueue,
   scoper: scoperQueue,
   multimd: multimdQueue,
-  pdb2crd: pdb2crdQueue,
   delete: deleteBilboMDJobsQueue
 }