+++ apps/worker/src/services/functions/mongo-utils.ts
@@ -6,7 +6,7 @@ import {
   IBilboMDSteps
 } from '@bilbomd/mongodb-schema'
 import { logger } from '../../helpers/loggers.js'
-
+type JobStatusEnum = 'Submitted' | 'Pending' | 'Running' | 'Completed' | 'Error'
 const updateStepStatus = async (
   job: IJob | IMultiJob,
   stepName: keyof IBilboMDSteps,
@@ -46,4 +46,12 @@ const handleStepError = async (
   logger.error(`Error in ${stepName}: ${errorMessage}`)
 }
 
-export { updateStepStatus, handleStepError }
+const updateJobStatus = async (
+  job: IJob,
+  status: JobStatusEnum
+): Promise<void> => {
+  job.status = status
+  await job.save()
+}
+
+export { updateStepStatus, handleStepError, updateJobStatus }