+++ apps/ui/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @bilbomd/ui
 
+## 1.24.2
+
+### Patch Changes
+
+- ef1f18f: Add simple component to display `md_constraints` if present
+  Adjust CSS FlexBox for SingleJobPage
+  Adjust CSS for cluster checkboxes in PAE Jiffy
+
+## 1.24.1
+
+### Patch Changes
+
+- 3e2f5b4: Update pnpm and dependencies
+
 ## 1.24.0
 
 ### Minor Changes